Update bluetooth-low-energy.service.ts
Add ability to use BLE device name. with IOS you can change the name with "nRF Connect" app. And it remains changed even after closing the app and locking the phone. You do need to restart the app and advertiser after a reboot of the device. but ill take it i tested for 24 hours and still had it working