ci: split linting into a separate job
Running the linting as part of the matrix previously ended up adding PR
comments twice for each linter issue on GitHub. As the codestyle doesn't
depend on the Node version we can save ourselves the duplicates this
way. Also prevents duplicate CI jobs when creating PRs from the same
repository.