feat(bluetooth-low-energy): add RSSI and measuredPower to entity
Tuning BLE values is tricky if the measured RSSI and the measuredPower
used for calculation are not visible anywhere. Putting them into the
attributes would cause a lot of data to be constantly updated in the
other integrations, so now they are just available in the API as an
additional field of the entity.

Closes #140, #177 and #203