feat(bluetooth-classic): add scan time limit option
Scanning for devices that aren't in range stresses the 2.4 GHz spectrum
for quite a while. With default settings it also barely left any time
for the integrated chips in Pis to communicate over WiFi. Resetting the
scan earlier should not have a notable impact on functionality for most
while hopefully decreasing the network performance hit.

Closes #179