refactor(xiaomi-mi): unlock distributed state
The use case for these entities being distributed is that the
devices
can be moved anyhwere in the house, with the current state still
being
reported as long as there is at least one room-assistant
instance
nearby. If the state is locked only the leader instance will
communicate
its updates to Home Assistant however. This ensures that any
node can
update the shared entities.