refactor(entities): emit all entity changes internally
As preparation for a generic MQTT integration the interface between
publishers and the entity registry was changed. Instead of only emitting
state and attribute events, all changes to entity properties are
emitted. The changes are now also always emitted, no matter if the
instance has authority over a distributed entity or not. The updates
contain an authority flag that should be checked by integrations that
need it.