feat: mark entities as unavailable on disconnect, shutdown & crash
This implementation makes the HA availability a bit more robust. In
comparison to the previous version it makes use of the last will message
to also mark entities available if the connection is lost or the
application doesn't exit gracefully. It also extends availability
support to distributed entities with the exception of the state-unlocked
ones.

Closes #181, closes #348