fix(home-assistant): use consistent id for instance availability
We need to match the behavior of the ClusterService to make it work
correctly across the cluster. Apart from that the instance name could
technically also contain characters that aren't valid for MQTT topic
names, so turning it into an ID like we do everywhere else isn't a bad
idea in general. For more info check the later posts in #728.